+++ test.conf
@@ -1,0 +1,4 @@
+server
+{
+listen 127.0.0.1:1234;
+}
